x/eightball/client/cli/query_kill_meditation_summoning.go
package cli

import (
	"context"
	"strconv"

	"eightball/x/eightball/types"
	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
)

func CmdListKillMeditationSummoning() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-kill-meditation-summoning",
		Short: "list all kill-meditation-summoning",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllKillMeditationSummoningRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.KillMeditationSummoningAll(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowKillMeditationSummoning() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-kill-meditation-summoning [id]",
		Short: "shows a kill-meditation-summoning",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			id, err := strconv.ParseUint(args[0], 10, 64)
			if err != nil {
				return err
			}

			params := &types.QueryGetKillMeditationSummoningRequest{
				Id: id,
			}

			res, err := queryClient.KillMeditationSummoning(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
